Counts and sorts the # of files within folders
This simple script counts the number of files within all the directories
that are located within a given path. Then it sorts the number of files
per folder by descending order.

I made this for some server software although it can be generally
useful, as well as a good example for beginner programmers. It shows how
you can use list comprehensions, converting / sorting dictionaries as
lists, and starting to use the os, and operator modules.